+++ src/Soluble/Metadata/Exception/UnexistentColumnException.php
@@ -1,0 +1,11 @@
+<?php
+
+/**
+ * @author Vanvelthem Sébastien
+ */
+
+namespace Soluble\Metadata\Exception;
+
+class UnexistentColumnException extends \RuntimeException
+{
+}