+++ test/SolubleTestFactories.php
@@ -1,11 +1,13 @@
 <?php
 
-class SolubleTestFactories {
+class SolubleTestFactories
+{
 
     /**
      * @return string
      */
-    public static function getCachePath() {
+    public static function getCachePath()
+    {
         $cache_dir = $_SERVER['PHPUNIT_CACHE_DIR'];
         if (!preg_match('/^\//', $cache_dir)) {
             $cache_dir = dirname(__FILE__) . DIRECTORY_SEPARATOR . $cache_dir;
@@ -20,7 +22,8 @@ class SolubleTestFactories {
      * @param string $charset
      * @return PDO|Mysqli|mixed
      */
-    public static function getDbConnection($type, $config = null, $charset = "UTF8") {
+    public static function getDbConnection($type, $config = null, $charset = "UTF8")
+    {
         if ($config === null) {
             $config = self::getDbConfiguration($type);
         }
@@ -30,9 +33,9 @@ class SolubleTestFactories {
         $database = isset($config['database']) ? $config['database'] : null;
         switch ($type) {
             case 'pdo:mysql':
-                $options = array(
+                $options = [
                     PDO::MYSQL_ATTR_INIT_COMMAND => "SET NAMES $charset",
-                );
+                ];
                 if ($database === null) {
                     $conn = new PDO("mysql:host=$hostname", $username, $password, $options);
                 } else {
@@ -64,14 +67,14 @@ class SolubleTestFactories {
 
             case 'doctrine2-mysqli' :
                 $params = \SolubleTestFactories::getDbConfiguration('mysqli');
-                $connectionParams = array(
+                $connectionParams = [
                     'dbname' => $params['database'],
                     'user' => $params['username'],
                     'password' => $params['password'],
                     'host' => $params['hostname'],
                     'driver' => 'mysqli',
                     'charset' => $charset
-                );
+                ];
                 $c = new \Doctrine\DBAL\Configuration();
                 $e = new \Doctrine\Common\EventManager();
                 //$r = new \ReflectionClass('Doctrine\Common\EventManager');
@@ -82,7 +85,7 @@ class SolubleTestFactories {
                 break;
             case 'zend-db2-mysqli' :
                 $params = \SolubleTestFactories::getDbConfiguration('mysqli');
-                $params = array_merge($params, array('driver' => 'Mysqli', 'charset' => $charset));
+                $params = array_merge($params, ['driver' => 'Mysqli', 'charset' => $charset]);
                 $conn = new \Zend\Db\Adapter\Adapter($params);
                 break;
             default:
@@ -96,19 +99,20 @@ class SolubleTestFactories {
      * @param string $type
      * @return array
      */
-    public static function getDbConfiguration($type) {
+    public static function getDbConfiguration($type)
+    {
         switch ($type) {
             case 'mysqli':
             case 'zend-db2-mysqli':
             case 'doctrine2-mysqli':
             case 'capsule5-mysqli':
             case 'pdo:mysql' :
-                $config = array(
+                $config = [
                     'hostname' => $_SERVER['MYSQL_HOSTNAME'],
                     'username' => $_SERVER['MYSQL_USERNAME'],
                     'password' => $_SERVER['MYSQL_PASSWORD'],
                     'database' => $_SERVER['MYSQL_DATABASE']
-                );
+                ];
                 break;
             default:
                 throw new \Exception("Database type unsupported in test configuration ($type)");
@@ -116,7 +120,8 @@ class SolubleTestFactories {
         return $config;
     }
 
-    public static function getDatabaseName($type) {
+    public static function getDatabaseName($type)
+    {
         $name = false;
         switch ($type) {
             case 'pdo:mysql':
@@ -128,5 +133,4 @@ class SolubleTestFactories {
         }
         return $name;
     }
-
 }