+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
-### 1.0.0 (2016--xx-xx)
+### 0.9.6 (2016-05-14)
+
+    - Updated README.md
+    - Added method `getColumn($name)` on `ColumnsMetadata`.
+    - More tests for differences between pdo_mysql and mysqli